examples/dpy_base_v2/requirements.txt
discord.py==2.2.3